[REFACTOR][mumi]: StoreLikeService, StoreLikeServiceImpl ->  StoreLikeCommandService, StoreLikeQueryService로 통합
